include property code in model; tests fix
