+++ src/main/java/gov/nih/nci/evs/api/service/ElasticSearchService.java
@@ -15,14 +15,6 @@ import gov.nih.nci.evs.api.model.SearchCriteria;
 public interface ElasticSearchService {
 
   /**
-   * Inits the settings.
-   *
-   * @throws IOException Signals that an I/O exception has occurred.
-   * @throws HttpClientErrorException the http client error exception
-   */
-  public void initSettings() throws IOException, HttpClientErrorException;
-
-  /**
    * Search.
    *
    * @param searchCriteria the search criteria
@@ -32,5 +24,4 @@ public interface ElasticSearchService {
    */
   public ConceptResultList search(SearchCriteria searchCriteria)
     throws IOException, HttpClientErrorException;
-
 }
