+++ src/test/java/gov/nih/nci/evs/api/controller/SearchControllerTests.java
@@ -405,57 +405,55 @@ public class SearchControllerTests {
     assertThat(content).isEqualTo(content2);
 
     // With property code also - P319
-    //TODO: test to be re-enabled after loading properties with code
-//    url = baseUrl;
-//    log.info("Testing url - " + url
-//        + "?terminology=ncit&term=XAV05295I5&property=P319&include=properties");
-//
-//    result = mvc
-//        .perform(get(url).param("terminology", "ncit").param("include", "properties")
-//            .param("term", "XAV05295I5").param("property", "P319"))
-//        .andExpect(status().isOk()).andReturn();
-//    content = result.getResponse().getContentAsString();
-//    log.info("  content = " + content);
-//    assertThat(content).isNotNull();
-//    list = new ObjectMapper().readValue(content, ConceptResultList.class);
-//    assertThat(list.getConcepts()).isNotNull();
-//    assertThat(list.getConcepts().size()).isEqualTo(1);
-//    assertThat(list.getConcepts().get(0).getName()).isEqualTo("Sivifene");
-//    assertThat(list.getConcepts().get(0).getProperties().stream()
-//        .filter(p -> p.getType().equals("FDA_UNII_Code")).count()).isGreaterThan(0);
-//    assertThat(list.getConcepts().get(0).getProperties().stream()
-//        .filter(p -> p.getType().equals("FDA_UNII_Code")).findFirst().get().getValue())
-//            .isEqualTo("XAV05295I5");
-
-    // Test with single terminology form
-    //TODO: test to be re-enabled after loading properties with code
-//    url = "/api/v1/concept/ncit/search";
-//    log.info("Testing url - " + url + "?term=XAV05295I5&property=P319&include=properties");
-//
-//    result = this.mvc.perform(get(url).param("include", "properties").param("term", "XAV05295I5")
-//        .param("property", "P319")).andExpect(status().isOk()).andReturn();
-//    content2 = result.getResponse().getContentAsString();
-//    log.info("content2 -" + content2);
-//    
-//    //removing timeTaken key from json before comparison
-//    content = removeTimeTaken(content);
-//    content2 = removeTimeTaken(content2);
-//    
-//    assertThat(content).isEqualTo(content2);
-
-    // BAD property type
     url = baseUrl;
-    log.info("Testing url - " + url + "?terminology=ncit&term=XAV05295I5&property=P999999");
+    log.info("Testing url - " + url
+        + "?terminology=ncit&term=XAV05295I5&property=P319&include=properties");
 
-    result = mvc.perform(get(url).param("terminology", "ncit").param("term", "XAV05295I5")
-        .param("property", "P999999")).andExpect(status().isBadRequest()).andReturn();
+    result = mvc
+        .perform(get(url).param("terminology", "ncit").param("include", "properties")
+            .param("term", "XAV05295I5").param("property", "P319"))
+        .andExpect(status().isOk()).andReturn();
+    content = result.getResponse().getContentAsString();
+    log.info("  content = " + content);
+    assertThat(content).isNotNull();
+    list = new ObjectMapper().readValue(content, ConceptResultList.class);
+    assertThat(list.getConcepts()).isNotNull();
+    assertThat(list.getConcepts().size()).isGreaterThan(0);
+    assertThat(list.getConcepts().get(0).getName()).isEqualTo("Sivifene");
+    assertThat(list.getConcepts().get(0).getProperties().stream()
+        .filter(p -> p.getType().equals("FDA_UNII_Code")).count()).isGreaterThan(0);
+    assertThat(list.getConcepts().get(0).getProperties().stream()
+        .filter(p -> p.getType().equals("FDA_UNII_Code")).findFirst().get().getValue())
+            .isEqualTo("XAV05295I5");
 
     // Test with single terminology form
     url = "/api/v1/concept/ncit/search";
-    log.info("Testing url - " + url + "?term=XAV05295I5&property=P999999");
+    log.info("Testing url - " + url + "?term=XAV05295I5&property=P319&include=properties");
 
-    result = this.mvc.perform(get(url).param("term", "XAV05295I5").param("property", "P999999"))
-        .andExpect(status().isBadRequest()).andReturn();
+    result = this.mvc.perform(get(url).param("include", "properties").param("term", "XAV05295I5")
+        .param("property", "P319")).andExpect(status().isOk()).andReturn();
+    content2 = result.getResponse().getContentAsString();
+    log.info("content2 -" + content2);
+    
+    //removing timeTaken key from json before comparison
+    content = removeTimeTaken(content);
+    content2 = removeTimeTaken(content2);
+    
+    assertThat(content).isEqualTo(content2);
+
+    // BAD property type
+//    url = baseUrl;
+//    log.info("Testing url - " + url + "?terminology=ncit&term=XAV05295I5&property=P999999");
+//
+//    result = mvc.perform(get(url).param("terminology", "ncit").param("term", "XAV05295I5")
+//        .param("property", "P999999")).andExpect(status().isBadRequest()).andReturn();
+//
+//    // Test with single terminology form
+//    url = "/api/v1/concept/ncit/search";
+//    log.info("Testing url - " + url + "?term=XAV05295I5&property=P999999");
+//
+//    result = this.mvc.perform(get(url).param("term", "XAV05295I5").param("property", "P999999"))
+//        .andExpect(status().isBadRequest()).andReturn();
 
     log.info("Done Testing testSearchProperty ");
   }
